Don't try merging with incompatible held item, with read-only inventory
Affects voxel-inventory-dialog. The result window is read-only, so you can
click with a held item to merge the result into it. However the merge
would still be attempted and the 'pickup' event sent even if the items could
not be merged, causing the result pile to be lost when you click it with
a different item. To fix this, check canPileWith() and return immediately.